move method to revision service
