+++ src/controllers/dimension.ts
@@ -29,7 +29,7 @@ import { createAllCubeFiles } from '../services/cube-builder';
 import { getFileService } from '../utils/get-file-service';
 import { TempFile } from '../interfaces/temp-file';
 import { cleanupTmpFile, uploadAvScan } from '../services/virus-scanner';
-import { updateRevisionTasks } from '../utils/update-revision-tasks';
+import { updateRevisionTasks } from '../services/revision';
 import { CubeBuildType } from '../enums/cube-build-type';
 import { BuildLog } from '../entities/dataset/build-log';
